Remove type: module form pacage.json Change extension for valalint files
