+++ package.json
@@ -1,11 +1,10 @@
 {
   "name": "@scality/core-ui",
-  "version": "0.206.0",
+  "version": "0.207.0",
   "description": "Scality common React component library",
   "author": "Scality Engineering",
   "license": "SEE LICENSE IN LICENSE",
   "main": "dist/index.js",
-  "type": "module",
   "types": "dist/index.d.ts",
   "mainSrc": "src/lib/index.js",
   "exports": {
