+++ src/lib/valalint/rules/modal-button-forbidden-label.test.mjs
