+++ src/lib/valalint/rules/technical-sentence-case.mjs
