+++ src/lib/valalint/rules/technical-sentence-case.test.mjs
